pyproject.toml
[tool.poetry]
name = "choosr"
version = "0.1.0"
description = ""
authors = ["Andy Clegg <andrew.clegg@pacerevenue.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.9"
tldextract = "^3.5.0"
pyyaml = "^6.0.2"

[tool.poetry.scripts]
'choosr' = 'choosr:main'


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
